dasda

+++ mmdet/models/dense_heads/detr_head.py
@@ -304,7 +304,7 @@ class DETRHead(BaseModule):
             scores[pos_inds] = bbox_overlaps(
                 pos_decode_bbox_pred.detach(),
                 pos_decode_bbox_targets,
-                is_aligned=True)
+                is_aligned=True).to(scores.dtype)
             loss_cls = self.loss_cls(
                 cls_scores, (labels, scores),
                 label_weights,
